+++ package/src/firebase_auth/index.ts
@@ -9,7 +9,7 @@ export { default as FirebaseAuthServerProvider } from './server/auth_user_server
 export { default as useFirebaseAuthUserClient } from './client/use_auth_user';
 export { default as useFirebaseAuthUserServer } from './server/use_auth_user_server';
 export { createLoginAction, createSignUpAction, createForgotPasswordAction } from './client/auth_actions';
-export { default as updateFirebaseAuthSession, sessionCookieName as firebaseAuthSessionCookieName } from './middleware/update_session';
+export { default as updateFirebaseAuthSession, defaultSessionCookieName as firebaseAuthSessionCookieName } from './middleware/update_session';
 export { getFirebaseAuthClient } from './client/firebase_client';
 export type { SerializedAuthUser, AuthFormState, AuthActionMessages, AuthUser } from './types';
 export type { FirebaseAuthRoutingConfig } from '../types/types';